+++ src/impulso/__init__.py
@@ -1,6 +1,7 @@
 from .acdc import Solver_ACDC, solve_ac, solve_dc, Statistics, Statistics
 from .base import TopologyError, Analysis
 from .circuit import Circuit
+from .components.array import Array
 from .components.bjt import BJT, NPN, PNP, emitter_current, collector_current, base_current
 from .components.capacitor import Capacitor
 from .sources.cccs import CCCS